Move to shared files.
No-Issue
